+++ quantus/functions/normalise_func.py
@@ -15,7 +15,6 @@ __all__ = [
     "normalise_by_max",
     "normalise_by_negative",
     "denormalise",
-    "normalize_sum_to_1",
     "normalise_by_average_second_moment_estimate",
 ]
 
@@ -256,12 +255,3 @@ def normalise_by_average_second_moment_estimate(
         )
 
     return a
-
-
-def normalize_sum_to_1(scores: np.ndarray, **kwargs) -> np.ndarray:
-    """Makes the absolute values sum to 1."""
-    if scores.ndim > 2:
-        raise ValueError("Only 2D and 1D inputs are supported.")
-    scores = np.asarray(scores)
-    scores = scores + np.finfo(np.float32).eps
-    return (scores.T / np.abs(scores).sum(axis=-1)).T
