+++ quantus/helpers/utils.py
@@ -21,6 +21,7 @@ from typing import (
     Union,
     List,
 )
+from functools import singledispatch
 
 import numpy as np
 from skimage.segmentation import slic, felzenszwalb
@@ -1076,8 +1077,32 @@ def get_logits_for_labels(logits: np.ndarray, y_batch: np.ndarray) -> np.ndarray
     return np.asarray(logits)[np.arange(y_batch.shape[0]), y_batch]
 
 
+@singledispatch
 def flatten_over_batch(arr: np.ndarray) -> np.ndarray:
     """Apply np.reshape(-1) keeping batch axis unchanged."""
     batch_size = len(arr)
     new_shape = (batch_size, -1)
     return np.reshape(arr, new_shape)
+
+
+@singledispatch
+def flatten_over_axis(arr: np.ndarray, axis: tuple[int, ...]) -> np.ndarray:
+    old_shape = np.take(np.shape(arr), axis)
+    new_shape = (*old_shape, -1)
+    return np.reshape(arr, new_shape)
+
+
+if is_tensorflow_available():
+    # To keep as much computations on GPU as possible.
+    @flatten_over_batch.register
+    def _(arr: tf.Tensor) -> tf.Tensor:
+        """Apply np.reshape(-1) keeping batch axis unchanged."""
+        batch_size = tf.shape(arr)[0]
+        new_shape = [batch_size, -1]
+        return tf.reshape(arr, new_shape)
+
+    @flatten_over_axis.register
+    def flatten_over_axis(arr: tf.Tensor, axis: tuple[int, ...]) -> tf.Tensor:
+        old_shape = tf.gather(tf.shape(arr), axis)
+        new_shape = tf.concat([old_shape, [-1]], axis=0)
+        return tf.reshape(arr, new_shape)
