+++ quantus/metrics/robustness/__init__.py
@@ -1,10 +1,15 @@
-from quantus.metrics.robustness.avg_sensitivity import AvgSensitivity
+from quantus.metrics.robustness.sensitivity import AvgSensitivity
 from quantus.metrics.robustness.consistency import Consistency
 from quantus.metrics.robustness.continuity import Continuity
 from quantus.metrics.robustness.local_lipschitz_estimate import LocalLipschitzEstimate
-from quantus.metrics.robustness.max_sensitivity import MaxSensitivity
-from quantus.metrics.robustness.relative_input_stability import RelativeInputStability
-from quantus.metrics.robustness.relative_output_stability import RelativeOutputStability
-from quantus.metrics.robustness.relative_representation_stability import (
+from quantus.metrics.robustness.sensitivity import (
+    MaxSensitivity,
+    AvgSensitivity,
+    AvgAndMaxSensitivity,
+)
+from quantus.metrics.robustness.relative_stability import (
+    RelativeInputStability,
+    RelativeOutputStability,
     RelativeRepresentationStability,
+    CombinedRelativeStability,
 )
