+++ quantus/metrics/robustness/sensitivity.py
@@ -7,8 +7,9 @@
 # Quantus project URL: <https://github.com/understandable-machine-intelligence-lab/Quantus>.
 from __future__ import annotations
 
+from abc import abstractmethod
 from functools import singledispatchmethod
-from typing import Any, Callable, Dict, List, Optional
+from typing import Any, Callable, Dict, List
 
 import numpy as np
 
@@ -16,10 +17,9 @@ from quantus.functions.norm_func import fro_norm
 from quantus.functions.normalise_func import normalise_by_max
 from quantus.functions.perturb_func import uniform_noise
 from quantus.functions.similarity_func import difference
-from quantus.helpers import asserts
-from quantus.helpers import warn
-from quantus.helpers.model.text_classifier import TextClassifier, Tokenizable
+from quantus.helpers import asserts, warn, collection_utils, utils, nlp_utils
 from quantus.helpers.model.model_interface import ModelInterface
+from quantus.helpers.model.text_classifier import TextClassifier, Tokenizable
 from quantus.helpers.q_types import (
     Explanation,
     ExplainFn,
@@ -30,15 +30,10 @@ from quantus.helpers.q_types import (
     NormFn,
     DataDomain,
 )
-from quantus.helpers.nlp_utils import (
-    is_plain_text_perturbation,
-    get_scores,
-)
-from quantus.helpers.collection_utils import value_or_default, add_default_items
 from quantus.metrics.base_batched import BatchedPerturbationMetric
 
 
-class AvgSensitivity(BatchedPerturbationMetric):
+class Sensitivity(BatchedPerturbationMetric):
     """
     Implementation of Avg-Sensitivity by Yeh at el., 2019.
 
@@ -124,8 +119,10 @@ class AvgSensitivity(BatchedPerturbationMetric):
         kwargs: optional
             Keyword arguments.
         """
-        perturb_func_kwargs = value_or_default(perturb_func_kwargs, lambda: {})
-        perturb_func_kwargs = add_default_items(
+        perturb_func_kwargs = collection_utils.value_or_default(
+            perturb_func_kwargs, lambda: {}
+        )
+        perturb_func_kwargs = collection_utils.add_default_items(
             perturb_func_kwargs, dict(lower_bound=lower_bound, upper_bound=upper_bound)
         )
 
@@ -328,21 +325,19 @@ class AvgSensitivity(BatchedPerturbationMetric):
             a_perturbed = self.explain_batch(model, x_batch, y_batch)
 
             # Measure similarity for each instance separately.
-            for instance_id in range(batch_size):
-                if instance_id in changed_prediction_indices:
-                    similarities[instance_id, step_id] = np.nan
-                    continue
+            sensitivities = self.similarity_func(
+                utils.flatten_over_axis(a_batch, (0, 1)),
+                utils.flatten_over_axis(a_perturbed, (0, 1)),
+            )
+            numerator = self.norm_numerator(sensitivities)
+            denominator = self.norm_denominator(utils.flatten_over_batch(x_batch))
+            sensitivities_norm = numerator / denominator
+            similarities[step_id] = sensitivities_norm
 
-                sensitivities = self.similarity_func(
-                    a_batch[instance_id].flatten(),
-                    a_perturbed[instance_id].flatten(),
-                )
-                numerator = self.norm_numerator(sensitivities)
-                denominator = self.norm_denominator(a=x_batch[instance_id].flatten())
-                sensitivities_norm = numerator / denominator
-                similarities[instance_id, step_id] = sensitivities_norm
-        mean_func = np.mean if self.return_nan_when_prediction_changes else np.nanmean
-        return mean_func(similarities, axis=1)
+            for i in changed_prediction_indices:
+                similarities[step_id, i] = np.nan
+
+        return self.reduce_similarities(similarities)
 
     @evaluate_batch.register
     def _(
@@ -356,7 +351,7 @@ class AvgSensitivity(BatchedPerturbationMetric):
     ) -> np.ndarray:
         batch_size = len(x_batch)
         similarities = np.zeros((batch_size, self.nr_samples)) * np.nan
-        is_plain_text = is_plain_text_perturbation(self.perturb_func)
+        is_plain_text = nlp_utils.is_plain_text_perturbation(self.perturb_func)
 
         for step_id in range(self.nr_samples):
             if is_plain_text:
@@ -368,8 +363,7 @@ class AvgSensitivity(BatchedPerturbationMetric):
                     model, x_batch, y_batch, a_batch
                 )
 
-        mean_func = np.mean if self.return_nan_when_prediction_changes else np.nanmean
-        return mean_func(similarities, axis=1)
+        return self.reduce_similarities(similarities)
 
     def _eval_step_nlp_plain_text(
         self,
@@ -379,11 +373,9 @@ class AvgSensitivity(BatchedPerturbationMetric):
         a_batch: list[Explanation],
         x_perturbed: list[str],
     ) -> np.ndarray:
-        batch_size = len(x_batch)
-        similarities = np.zeros(shape=batch_size)
-        changed_prediction_indices = self.changed_prediction_indices(
-            model, x_batch, x_perturbed
-        )
+        # changed_prediction_indices = self.changed_prediction_indices(
+        #    model, x_batch, x_perturbed
+        # )
 
         for x_instance, x_instance_perturbed in zip(x_batch, x_perturbed):
             warn.warn_perturbation_caused_no_change(
@@ -393,29 +385,22 @@ class AvgSensitivity(BatchedPerturbationMetric):
         # Generate explanation based on perturbed input x.
         a_perturbed = self.explain_batch(model, x_perturbed, y_batch)
         # Get numerical part of explanations.
-        a_batch = get_scores(a_batch)
-        a_perturbed = get_scores(a_perturbed)
+        a_batch = nlp_utils.get_scores(a_batch)
+        a_perturbed = nlp_utils.get_scores(a_perturbed)
 
         # Get numerical representation of x_batch.
         x_batch_embeddings, predict_kwargs = model.get_embeddings(x_batch)
 
-        # Measure similarity for each instance separately.
-        for instance_id in range(batch_size):
-            if instance_id in changed_prediction_indices:
-                similarities[instance_id] = np.nan
-                continue
+        sensitivities = self.similarity_func(
+            utils.flatten_over_axis(a_batch, (0, 1)),
+            utils.flatten_over_axis(a_perturbed, (0, 1)),
+        )
 
-            sensitivities = self.similarity_func(
-                a=np.reshape(a_batch[instance_id], -1),
-                b=np.reshape(a_perturbed[instance_id], -1),
-            )
-            numerator = self.norm_numerator(a=sensitivities)
-            denominator = self.norm_denominator(
-                a=np.reshape(x_batch_embeddings[instance_id], -1)
-            )
-            sensitivities_norm = numerator / denominator
-            similarities[instance_id] = sensitivities_norm
-
+        numerator = self.norm_numerator(sensitivities)
+        denominator = self.norm_denominator(
+            utils.flatten_over_batch(x_batch_embeddings)
+        )
+        similarities = numerator / denominator
         return similarities
 
     def _eval_step_nlp_embeddings(
@@ -425,17 +410,14 @@ class AvgSensitivity(BatchedPerturbationMetric):
         y_batch: np.ndarray,
         a_batch: list[Explanation],
     ) -> np.ndarray:
-        batch_size = len(x_batch)
-        similarities = np.zeros(shape=batch_size)
-
         x_batch_embeddings, predict_kwargs = model.get_embeddings(x_batch)
-        a_batch = get_scores(a_batch)
+        a_batch = nlp_utils.get_scores(a_batch)
 
         # Perturb input.
         x_perturbed = self.perturb_batch(x_batch_embeddings)
-        changed_prediction_indices = self.changed_prediction_indices(
-            model, x_batch_embeddings, x_perturbed
-        )
+        # changed_prediction_indices = self.changed_prediction_indices(
+        #   model, x_batch_embeddings, x_perturbed
+        #
 
         for x_instance, x_instance_perturbed in zip(x_batch_embeddings, x_perturbed):
             warn.warn_perturbation_caused_no_change(
@@ -445,23 +427,15 @@ class AvgSensitivity(BatchedPerturbationMetric):
         # Generate explanation based on perturbed input x.
         a_perturbed = self.explain_batch(model, x_perturbed, y_batch, **predict_kwargs)
 
-        # Measure similarity for each instance separately.
-        for instance_id in range(batch_size):
-            if instance_id in changed_prediction_indices:
-                similarities[instance_id] = np.nan
-                continue
-
-            sensitivities = self.similarity_func(
-                np.reshape(a_batch[instance_id], -1),
-                np.reshape(a_perturbed[instance_id], -1),
-            )
-            numerator = self.norm_numerator(sensitivities)
-            denominator = self.norm_denominator(
-                np.reshape(x_batch_embeddings[instance_id], -1)
-            )
-            sensitivities_norm = numerator / denominator
-            similarities[instance_id] = sensitivities_norm
-
+        sensitivities = self.similarity_func(
+            utils.flatten_over_axis(a_batch, (0, 1)),
+            utils.flatten_over_axis(a_perturbed, (0, 1)),
+        )
+        numerator = self.norm_numerator(sensitivities)
+        denominator = self.norm_denominator(
+            utils.flatten_over_batch(x_batch_embeddings)
+        )
+        similarities = numerator / denominator
         return similarities
 
     def custom_preprocess(
@@ -498,3 +472,270 @@ class AvgSensitivity(BatchedPerturbationMetric):
         # Additional explain_func assert, as the one in prepare() won't be
         # executed when a_batch != None.
         asserts.assert_explain_func(explain_func=self.explain_func)
+
+    @abstractmethod
+    def reduce_similarities(self, similarities: np.ndarray) -> np.ndarray:
+        raise NotImplementedError
+
+
+class MaxSensitivity(Sensitivity):
+    """
+    Implementation of Max-Sensitivity by Yeh at el., 2019.
+
+    Using Monte Carlo sampling-based approximation while measuring how explanations
+    change under slight perturbation - the maximum sensitivity is captured.
+
+    References:
+        1) Chih-Kuan Yeh et al. "On the (in) fidelity and sensitivity for explanations."
+        NeurIPS (2019): 10965-10976.
+        2) Umang Bhatt et al.: "Evaluating and aggregating
+        feature-based model explanations."  IJCAI (2020): 3016-3022.
+    """
+
+    def reduce_similarities(self, similarities: np.ndarray) -> np.ndarray:
+        max_func = np.max if self.return_nan_when_prediction_changes else np.nanmax
+        return max_func(similarities, axis=1)
+
+
+class AvgSensitivity(Sensitivity):
+    """
+    Implementation of Max-Sensitivity by Yeh at el., 2019.
+
+    Using Monte Carlo sampling-based approximation while measuring how explanations
+    change under slight perturbation - the maximum sensitivity is captured.
+
+    References:
+        1) Chih-Kuan Yeh et al. "On the (in) fidelity and sensitivity for explanations."
+        NeurIPS (2019): 10965-10976.
+        2) Umang Bhatt et al.: "Evaluating and aggregating
+        feature-based model explanations."  IJCAI (2020): 3016-3022.
+    """
+
+    def reduce_similarities(self, similarities: np.ndarray) -> np.ndarray:
+        mean_func = np.mean if self.return_nan_when_prediction_changes else np.nanmean
+        return mean_func(similarities, axis=1)
+
+
+class AvgAndMaxSensitivity(Sensitivity):
+    """
+    Computes average and maximal sensitivity with computational cost of one metric.
+    During large scale evaluation, it is often required to compute both.
+    Using them separately, means recomputing twice all perturbed explanations,
+    and for advanced XAI methods, this is THE MOST computationally expensive part.
+    """
+
+    @asserts.attributes_check
+    def __init__(
+        self,
+        similarity_func: SimilarityFn = difference,
+        norm_numerator: NormFn = fro_norm,
+        norm_denominator: NormFn = fro_norm,
+        nr_samples: int = 200,
+        abs: bool = False,
+        normalise: bool = False,
+        normalise_func: NormaliseFn = normalise_by_max,
+        normalise_func_kwargs: Dict[str, Any] | None = None,
+        perturb_func: PerturbFn = uniform_noise,
+        lower_bound: float = 0.2,
+        upper_bound: float | None = None,
+        perturb_func_kwargs: Dict[str, ...] | None = None,
+        return_aggregate: bool = False,
+        aggregate_func: AggregateFn = np.mean,
+        default_plot_func: Callable[[...], None] | None = None,
+        disable_warnings: bool = False,
+        display_progressbar: bool = False,
+        return_nan_when_prediction_changes: bool = False,
+        **kwargs,
+    ):
+        """
+        Parameters
+        ----------
+        similarity_func: callable
+            Similarity function applied to compare input and perturbed input.
+            If None, the default value is used, default=difference.
+        norm_numerator: callable
+            Function for norm calculations on the numerator.
+            If None, the default value is used, default=fro_norm
+        norm_denominator: callable
+            Function for norm calculations on the denominator.
+            If None, the default value is used, default=fro_norm
+        nr_samples: integer
+            The number of samples iterated, default=200.
+        normalise: boolean
+            Indicates whether normalise operation is applied on the attribution, default=True.
+        normalise_func: callable
+            Attribution normalisation function applied in case normalise=True.
+            If normalise_func=None, the default value is used, default=normalise_by_max.
+        normalise_func_kwargs: dict
+            Keyword arguments to be passed to normalise_func on call, default={}.
+        perturb_func: callable
+            Input perturbation function. If None, the default value is used,
+            default=gaussian_noise.
+        perturb_std: float
+            The amount of noise added, default=0.1.
+        perturb_mean: float
+            The mean of noise added, default=0.0.
+        perturb_func_kwargs: dict
+            Keyword arguments to be passed to perturb_func, default={}.
+        return_aggregate: boolean
+            Indicates if an aggregated score should be computed over all instances.
+        aggregate_func: callable
+            Callable that aggregates the scores given an evaluation call.
+        default_plot_func: callable
+            Callable that plots the metrics result.
+        disable_warnings: boolean
+            Indicates whether the warnings are printed, default=False.
+        display_progressbar: boolean
+            Indicates whether a tqdm-progress-bar is printed, default=False.
+        return_nan_when_prediction_changes: boolean
+            When set to true, the metric will be evaluated to NaN if the prediction changes after the perturbation is applied.
+        kwargs: optional
+            Keyword arguments.
+        """
+
+        super().__init__(
+            similarity_func,
+            norm_numerator,
+            norm_denominator,
+            nr_samples,
+            abs,
+            normalise,
+            normalise_func,
+            normalise_func_kwargs,
+            perturb_func,
+            lower_bound,
+            upper_bound,
+            perturb_func_kwargs,
+            return_aggregate,
+            aggregate_func,
+            default_plot_func,
+            disable_warnings,
+            display_progressbar,
+            return_nan_when_prediction_changes,
+        )
+
+    def __call__(
+        self,
+        model,
+        x_batch: np.array,
+        y_batch: np.array,
+        explain_func: ExplainFn,
+        a_batch: np.ndarray | None = None,
+        s_batch: np.ndarray | None = None,
+        channel_first: bool | None = None,
+        explain_func_kwargs: dict | None = None,
+        model_predict_kwargs: dict | None = None,
+        softmax: bool | None = False,
+        device: str | None = None,
+        batch_size: int = 64,
+        custom_batch: Any | None = None,
+        tokenizer: Tokenizable | None = None,
+        **kwargs,
+    ) -> np.ndarray | float:
+        """
+        This implementation represents the main logic of the metric and makes the class object callable.
+        It completes instance-wise evaluation of explanations (a_batch) with respect to input data (x_batch),
+        output labels (y_batch) and a torch or tensorflow model (model).
+
+        Calls general_preprocess() with all relevant arguments, calls
+        () on each instance, and saves results to last_results.
+        Calls custom_postprocess() afterwards. Finally returns last_results.
+
+        Parameters
+        ----------
+        model: torch.nn.Module, tf.keras.Model
+            A torch or tensorflow model that is subject to explanation.
+        x_batch: np.ndarray
+            A np.ndarray which contains the input data that are explained.
+        y_batch: np.ndarray
+            A np.ndarray which contains the output labels that are explained.
+        a_batch: np.ndarray, optional
+            A np.ndarray which contains pre-computed attributions i.e., explanations.
+        s_batch: np.ndarray, optional
+            A np.ndarray which contains segmentation masks that matches the input.
+        channel_first: boolean, optional
+            Indicates of the image dimensions are channel first, or channel last.
+            Inferred from the input shape if None.
+        explain_func: callable
+            Callable generating attributions.
+        explain_func_kwargs: dict, optional
+            Keyword arguments to be passed to explain_func on call.
+        model_predict_kwargs: dict, optional
+            Keyword arguments to be passed to the model's predict method.
+        softmax: boolean
+            Indicates whether to use softmax probabilities or logits in model prediction.
+            This is used for this __call__ only and won't be saved as attribute. If None, self.softmax is used.
+        device: string
+            Indicated the device on which a torch.Tensor is or will be allocated: "cpu" or "gpu".
+        kwargs: optional
+            Keyword arguments.
+
+        Returns
+        -------
+        last_results: list
+            a list of Any with the evaluation scores of the concerned batch.
+
+        Examples:
+        --------
+            # Minimal imports.
+            >> import quantus
+            >> from quantus import LeNet
+            >> import torch
+
+            # Enable GPU.
+            >> device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+
+            # Load a pre-trained LeNet classification model (architecture at quantus/helpers/models).
+            >> model = LeNet()
+            >> model.load_state_dict(torch.load("tutorials/assets/pytests/mnist_model"))
+
+            # Load MNIST datasets and make loaders.
+            >> test_set = torchvision.datasets.MNIST(root='./sample_data', download=True)
+            >> test_loader = torch.utils.data.DataLoader(test_set, batch_size=24)
+
+            # Load a batch of inputs and outputs to use for XAI evaluation.
+            >> x_batch, y_batch = iter(test_loader).next()
+            >> x_batch, y_batch = x_batch.cpu().numpy(), y_batch.cpu().numpy()
+
+            # Generate Saliency attributions of the test set batch of the test set.
+            >> a_batch_saliency = Saliency(model).attribute(inputs=x_batch, target=y_batch, abs=True).sum(axis=1)
+            >> a_batch_saliency = a_batch_saliency.cpu().numpy()
+
+            # Initialise the metric and evaluate explanations by calling the metric instance.
+            >> metric = Metric(abs=True, normalise=False)
+            >> scores = metric(model=model, x_batch=x_batch, y_batch=y_batch, a_batch=a_batch_saliency}
+        """
+        return super().__call__(
+            model=model,
+            x_batch=x_batch,
+            y_batch=y_batch,
+            a_batch=a_batch,
+            s_batch=s_batch,
+            custom_batch=None,
+            channel_first=channel_first,
+            explain_func=explain_func,
+            explain_func_kwargs=explain_func_kwargs,
+            softmax=softmax,
+            device=device,
+            model_predict_kwargs=model_predict_kwargs,
+            batch_size=batch_size,
+            tokenizer=tokenizer,
+            **kwargs,
+        )
+
+    def reduce_similarities(self, similarities: np.ndarray) -> dict[str, np.ndarray]:
+        mean_func = np.mean if self.return_nan_when_prediction_changes else np.nanmean
+        max_func = np.max if self.return_nan_when_prediction_changes else np.nanmax
+        avg = mean_func(similarities, axis=1)
+        _max = max_func(similarities, axis=1)
+        return dict(avg=avg, max=_max)
+
+    @staticmethod
+    def join_batches(
+        score_batches: List[dict[str, np.ndarray]]
+    ) -> dict[str, np.ndarray]:
+        scores = {"avg": [], "max": []}
+        for s in score_batches:
+            scores["avg"].extend(s["avg"])
+            scores["max"].extend(s["max"])
+        return collection_utils.map_dict(scores, np.asarray)
